requirements.txt
torch>=2.2
torchvision>=0.17
timm>=1.0.9
numpy>=1.24
pillow>=10.0
matplotlib>=3.7
scikit-learn>=1.3
tqdm>=4.66
pyyaml>=6.0
datasets>=2.19
pandas>=2.0

